modify editprofile, top-selling,administrator

+++ src/app/user/account/address-book/address-book.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from '@angular/core';
+import {ChangeDetectorRef, Component, OnInit} from '@angular/core';
 import {UserInfoService} from "../../../shared/service/user-info.service";
 import {UserInfo} from "../../../shared/model/UserInfo";
 import {ActivatedRoute} from "@angular/router";
@@ -7,6 +7,7 @@ import {DialogRef} from "@angular/cdk/dialog";
 import {AddressDialogComponent} from "./address-dialog/address-dialog.component";
 import {MatDialog} from "@angular/material/dialog";
 import {AddressBook} from "../../../shared/model/AddressBook";
+import {ConfirmDialogComponent} from "../../../shared/dialog/confirm-dialog.component";
 
 @Component({
   selector: 'app-address-book',
@@ -17,7 +18,8 @@ export class AddressBookComponent implements OnInit{
   userInfo$=new BehaviorSubject<UserInfo>({} as UserInfo);
   constructor(public userInfoService: UserInfoService,
               private route:ActivatedRoute,
-              private dialog: MatDialog) {
+              private dialog: MatDialog,
+              private cdr: ChangeDetectorRef) {
   }
 
   ngOnInit(): void {
@@ -51,12 +53,25 @@ export class AddressBookComponent implements OnInit{
   }
 
   deleteAddress(address: AddressBook) {
-    this.userInfoService.deleteAddress(this.userInfo$.value.id, address).subscribe(res=>{
-      if (res.success) {
-        this.userInfoService.userInfo = res.data;
-      } else {
-        console.log(res);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Confirm',
+        message: 'Are you sure to delete this address?'
       }
-    })
+    });
+    dialogRef.afterClosed().subscribe(res => {
+      if (res) {
+        this.userInfoService.deleteAddress(this.userInfo$.value.id, address).subscribe(res=>{
+          if (res.success) {
+            this.userInfoService.userInfo = res.data;
+            this.userInfo$.next(res.data);
+            this.cdr.detectChanges();
+          } else {
+            console.log(res);
+          }
+        })
+      }
+    });
+
   }
 }
